Refactor SdlSession2 into SdlSession with base
